+++ examples/basic-server/examples/users.rs
@@ -5,10 +5,12 @@ use std::time::Duration;
 use bonsaidb::{
     client::{url::Url, Client},
     core::{
-        admin::PermissionGroup,
+        admin::{PermissionGroup, Role},
         connection::{AsyncStorageConnection, Authentication, SensitiveString},
         permissions::{
-            bonsai::{AuthenticationMethod, BonsaiAction, ServerAction},
+            bonsai::{
+                AuthenticationMethod, BonsaiAction, DatabaseAction, DocumentAction, ServerAction,
+            },
             Permissions, Statement,
         },
         schema::{InsertError, SerializedCollection},
@@ -24,6 +26,28 @@ use support::schema::Shape;
 async fn main() -> anyhow::Result<()> {
     drop(env_logger::try_init());
     let server = setup_server().await?;
+
+    // This example shows off basic user authentication as well as the ability
+    // to assume roles. The server will be configured to only allow connected
+    // users the ability to authenticate. All other usage will result in
+    // PermissionDenied errors.
+    //
+    // We will create a user that belongs to a PermissionGroup that gives it the
+    // ability to insert and read documents, but not delete them. This example
+    // will demonstrate how to authenticate as the user, and shows how
+    // permission is denied before authentication but is allowed on the
+    // authenticated client.
+    //
+    // The other bit of setup we are going to do is create an administrators
+    // group that the user belongs to. This permission group will allow the user
+    // to assume any identity (user or role) in BonsaiDb. We are going to show
+    // how to use this to escalate privileges by creating a "superuser" Role,
+    // which belongs to a "superusers" group that grants all privileges.
+    //
+    // This example will finish by using the authenticated client to assume the
+    // Superuser role and delete the record we inserted. While this is a complex
+    // setup, it is a powerful pattern in Role Based Access Control which can
+    // help protect users from accidentally performing a dangerous operation.
 
     // Create a database user, or get its ID if it already existed.
     let user_id = match server.create_user("ecton").await {
@@ -41,11 +65,79 @@ async fn main() -> anyhow::Result<()> {
         Err(other) => anyhow::bail!(other),
     };
 
-    // Create an administrators permission group, or get its ID if it already existed.
+    // Create an basic-users permission group, or get its ID if it already existed.
     let admin = server.admin().await;
-    let administrator_group_id = match (PermissionGroup {
-        name: String::from("administrators"),
+    let users_group_id = match (PermissionGroup {
+        name: String::from("basic-users"),
+        statements: vec![Statement::for_any()
+            .allowing(&BonsaiAction::Database(DatabaseAction::Document(
+                DocumentAction::Insert,
+            )))
+            .allowing(&BonsaiAction::Database(DatabaseAction::Document(
+                DocumentAction::Get,
+            )))],
+    }
+    .push_into_async(&admin)
+    .await)
+    {
+        Ok(doc) => doc.header.id,
+        Err(InsertError {
+            error:
+                bonsaidb::core::Error::UniqueKeyViolation {
+                    existing_document, ..
+                },
+            ..
+        }) => existing_document.id.deserialize()?,
+        Err(other) => anyhow::bail!(other),
+    };
+
+    // Make our user a member of the basic-users group.
+    server
+        .add_permission_group_to_user(user_id, users_group_id)
+        .await?;
+
+    // Create an superusers group, which has all permissions
+    let superusers_group_id = match (PermissionGroup {
+        name: String::from("superusers"),
         statements: vec![Statement::allow_all_for_any_resource()],
+    }
+    .push_into_async(&admin)
+    .await)
+    {
+        Ok(doc) => doc.header.id,
+        Err(InsertError {
+            error:
+                bonsaidb::core::Error::UniqueKeyViolation {
+                    existing_document, ..
+                },
+            ..
+        }) => existing_document.id.deserialize()?,
+        Err(other) => anyhow::bail!(other),
+    };
+
+    let superuser_role_id = match (Role {
+        name: String::from("superuser"),
+        groups: vec![superusers_group_id],
+    }
+    .push_into_async(&admin)
+    .await)
+    {
+        Ok(doc) => doc.header.id,
+        Err(InsertError {
+            error:
+                bonsaidb::core::Error::UniqueKeyViolation {
+                    existing_document, ..
+                },
+            ..
+        }) => existing_document.id.deserialize()?,
+        Err(other) => anyhow::bail!(other),
+    };
+
+    let administrators_group_id = match (PermissionGroup {
+        name: String::from("administrators"),
+        statements: vec![
+            Statement::for_any().allowing(&BonsaiAction::Server(ServerAction::AssumeIdentity))
+        ],
     }
     .push_into_async(&admin)
     .await)
@@ -63,10 +155,8 @@ async fn main() -> anyhow::Result<()> {
 
     // Make our user a member of the administrators group.
     server
-        .add_permission_group_to_user(user_id, administrator_group_id)
+        .add_permission_group_to_user(user_id, administrators_group_id)
         .await?;
-
-    // ANCHOR_END: setup
 
     // Give a moment for the listeners to start.
     tokio::time::sleep(Duration::from_millis(10)).await;
@@ -101,13 +191,26 @@ async fn main() -> anyhow::Result<()> {
             "ecton",
             Authentication::Password(SensitiveString(String::from("hunter2"))),
         )
-        .await
-        .unwrap();
+        .await?;
     let db = authenticated_client
         .database::<Shape>("my-database")
         .await?;
     let shape_doc = Shape::new(3).push_into_async(&db).await?;
-    log::info!("Successully inserted document {:?}", shape_doc);
+    println!("Successully inserted document {:?}", shape_doc);
+
+    // The "basic-users" group and "administrators" groups do not give
+    // permission to delete documents:
+    assert!(matches!(
+        shape_doc.delete_async(&db).await.unwrap_err(),
+        bonsaidb::core::Error::PermissionDenied { .. }
+    ));
+
+    // But we can assume the Superuser role to delete the document:
+    let as_superuser =
+        Role::assume_identity_async(superuser_role_id, &authenticated_client).await?;
+    shape_doc
+        .delete_async(&as_superuser.database::<Shape>("my-database").await?)
+        .await?;
 
     drop(db);
     drop(client);